# docs/generated/openapi.rapidapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Naruto Mythos TCG Developer API",
    "description": "A developer-friendly REST API for querying the Naruto Mythos TCG catalogue. Card is a canonical identity; Printing is a collectible identity; Edition is scoped to a Set.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/editions/{edition_id}/printings": {
      "get": {
        "tags": [
          "Editions",
          "Collector Catalogue"
        ],
        "summary": "List an Edition's verified collector catalogue",
        "description": "Returns curated collector-reference identities when available. This intentionally differs from GET /v1/printings, which returns every persisted Printing including source-only rows.",
        "operationId": "edition_printings_v1_editions__edition_id__printings_get",
        "parameters": [
          {
            "name": "edition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Edition Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPrintingsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Edition not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EDITION_NOT_FOUND",
                    "message": "Edition not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or sort parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "pagination": {
                    "value": {
                      "error": {
                        "code": "INVALID_PAGINATION",
                        "message": "page must be >= 1."
                      }
                    }
                  },
                  "sort": {
                    "value": {
                      "error": {
                        "code": "INVALID_SORT",
                        "message": "Invalid sort field."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/printings": {
      "get": {
        "tags": [
          "Printings"
        ],
        "summary": "List persisted Printings",
        "description": "Returns every persisted collectible Printing, including source-only Printings not included in a verified collector reference. Card is the canonical identity; Printing is the collectible identity.",
        "operationId": "list_printings_v1_printings_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "set",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Set public ID.",
              "title": "Set"
            },
            "description": "Set public ID."
          },
          {
            "name": "edition",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Edition public ID.",
              "title": "Edition"
            },
            "description": "Edition public ID."
          },
          {
            "name": "rarity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Normalized rarity, case-insensitive.",
              "title": "Rarity"
            },
            "description": "Normalized rarity, case-insensitive."
          },
          {
            "name": "treatment",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Normalized Printing treatment, case-insensitive.",
              "title": "Treatment"
            },
            "description": "Normalized Printing treatment, case-insensitive."
          },
          {
            "name": "collector_class",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Normalized collector class.",
              "title": "Collector Class"
            },
            "description": "Normalized collector class."
          },
          {
            "name": "card",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Canonical Card public ID.",
              "title": "Card"
            },
            "description": "Canonical Card public ID."
          },
          {
            "name": "origin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "UPSTREAM, UPSTREAM_AND_CURATED, or CURATED_REFERENCE.",
              "title": "Origin"
            },
            "description": "UPSTREAM, UPSTREAM_AND_CURATED, or CURATED_REFERENCE."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPrintingsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/printings/{printing_id}": {
      "get": {
        "tags": [
          "Printings"
        ],
        "summary": "Get a Printing",
        "description": "Returns one persisted collectible Printing and derives its origin from upstream and curated-reference relationships.",
        "operationId": "get_printing_v1_printings__printing_id__get",
        "parameters": [
          {
            "name": "printing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Printing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintingResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardImageResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "example": "front"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "example": "https://example.invalid/test-card.jpg"
          },
          "width": {
            "type": "integer",
            "nullable": true,
            "title": "Width",
            "example": 734
          },
          "height": {
            "type": "integer",
            "nullable": true,
            "title": "Height",
            "example": 1024
          }
        },
        "type": "object",
        "required": [
          "type",
          "url"
        ],
        "title": "CardImageResponse",
        "example": {
          "height": 1024,
          "type": "front",
          "url": "https://example.invalid/test-card.jpg",
          "width": 734
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "CARD_NOT_FOUND",
          "SET_NOT_FOUND",
          "EDITION_NOT_FOUND",
          "PRINTING_NOT_FOUND",
          "COLLECTOR_REFERENCE_NOT_FOUND",
          "KEYWORD_NOT_FOUND",
          "INVALID_FILTER",
          "INVALID_PAGINATION",
          "INVALID_SORT",
          "FORBIDDEN",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "description": "Known/expected error codes for the /v1 API.\n\nThis list is expected to grow in later phases as route handlers are\nimplemented (e.g. Phase 5+ will need to actually raise CARD_NOT_FOUND\nwhen a lookup misses) -- growing it is an additive change."
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedPrintingsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PrintingResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "PaginatedPrintingsResponse"
      },
...
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "example": 50
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "example": 630
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages",
            "example": 13
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next",
            "example": true
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous",
            "example": false
          }
        },
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "pages",
          "has_next",
          "has_previous"
        ],
        "title": "PaginationMeta",
        "example": {
          "has_next": true,
          "has_previous": false,
          "limit": 50,
          "page": 1,
          "pages": 13,
          "total": 630
        }
      },
...
      "PrintingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "card_id": {
            "type": "string",
            "title": "Card Id"
          },
          "card_name": {
            "type": "string",
            "title": "Card Name"
          },
          "card_type": {
            "type": "string",
            "nullable": true,
            "title": "Card Type"
          },
          "set_id": {
            "type": "string",
            "title": "Set Id"
          },
          "set_name": {
            "type": "string",
            "title": "Set Name"
          },
          "edition_id": {
            "type": "string",
            "nullable": true,
            "title": "Edition Id"
          },
          "edition_name": {
            "type": "string",
            "nullable": true,
            "title": "Edition Name"
          },
          "collector_number": {
            "type": "string",
            "nullable": true,
            "title": "Collector Number"
          },
          "printed_number": {
            "type": "string",
            "title": "Printed Number"
          },
          "rarity": {
            "type": "string",
            "nullable": true,
            "title": "Rarity"
          },
          "collector_class": {
            "type": "string",
            "nullable": true,
            "title": "Collector Class"
          },
          "treatment": {
            "type": "string",
            "nullable": true,
            "title": "Treatment"
          },
          "finish": {
            "type": "string",
            "nullable": true,
            "title": "Finish"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "serial_numbered": {
            "type": "boolean",
            "title": "Serial Numbered"
          },
          "serial_total": {
            "type": "integer",
            "nullable": true,
            "title": "Serial Total"
          },
          "origin": {
            "type": "string",
            "pattern": "^(UPSTREAM|UPSTREAM_AND_CURATED|CURATED_REFERENCE)$",
            "title": "Origin"
          },
          "source_rarity": {
            "type": "string",
            "nullable": true,
            "title": "Source Rarity"
          },
          "source_variant": {
            "type": "string",
            "nullable": true,
            "title": "Source Variant"
          },
          "collector_reference_number": {
            "type": "string",
            "nullable": true,
            "title": "Collector Reference Number"
          },
          "collector_reference_card_name": {
            "type": "string",
            "nullable": true,
            "title": "Collector Reference Card Name"
          },
          "collector_reference_rarity": {
            "type": "string",
            "nullable": true,
            "title": "Collector Reference Rarity"
          },
          "collector_reference_variant": {
            "type": "string",
            "nullable": true,
            "title": "Collector Reference Variant"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/CardImageResponse"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_id",
          "card_name",
          "set_id",
          "set_name",
          "printed_number",
          "language",
          "serial_numbered",
          "origin"
        ],
        "title": "PrintingResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}